changed styles for skeleton
