+++ src/core/components/error-component/styles.js
@@ -2,7 +2,7 @@ const styles = (theme) => ({
   error: {
     background: "#37474f",
     color: "white",
-    margin: "2em auto",
+    margin: "0 auto",
     textTransform: "uppercase",
     height: "30rem",
   },
@@ -11,7 +11,7 @@ const styles = (theme) => ({
     color: "#fff",
     margin: "0.5em 2em",
     align: "justify",
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down("xs")]: {
       margin: "0 auto 1em",
       fontSize: "1.2em",
     },
@@ -19,8 +19,8 @@ const styles = (theme) => ({
   errorMistake: {
     margin: "5em 0 2em",
     fontSize: "2.4rem",
-    [theme.breakpoints.down("sm")]: {
-      margin: "0 auto 1em",
+    [theme.breakpoints.down("xs")]: {
+      margin: "1em auto 1em",
       fontSize: "1.7em",
     },
   },
