+++ src/core/components/skeleton/skeleton.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import Skeleton from "@material-ui/lab/Skeleton";
 import { Grid } from "@material-ui/core";
-import styles from "../../../styles";
+import styles from "./styles";
 import "./skeleton.scss";
 import { withStyles } from "@material-ui/core/styles";
